test: 테스트 Fixture 분리
