+++ src/test/java/racingcar/controller/RacingGameWebControllerTest.java
@@ -4,28 +4,16 @@ package racingcar.controller;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import io.restassured.specification.RequestSpecification;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import racingcar.controller.dto.GameRequest;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class RacingGameWebControllerTest {
+class RacingGameWebControllerTest extends RacingGameFixtures {
 
-    @BeforeEach
-    void setUp(@LocalServerPort final int port) {
-        RestAssured.port = port;
-    }
 
     @DisplayName("doGame을 실행하면 게임의 결과가 반환된다.")
     @Test
@@ -64,20 +52,5 @@ class RacingGameWebControllerTest {
                 () -> assertThat(result.getList("racingCars")).hasSize(2),
                 () -> assertThat(result.getList("racingCars.name")).containsExactly(List.of("콩하나", "에단"), List.of("준팍", "에코", "소니"))
         );
-    }
-
-    private static ExtractableResponse<Response> savePlayers(String names, int trialCount) {
-        final GameRequest request = new GameRequest(names, trialCount);
-
-        return given()
-                .body(request)
-                .when().post("/plays")
-                .then().log().all()
-                .extract();
-    }
-
-    private static RequestSpecification given() {
-        return RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE);
     }
 }
